add hash table

+++ src/server/socket_pkg.h
@@ -40,7 +40,7 @@ typedef struct socket_pkg{
 	uint32_t           data_len;     //数据长度
 	uint16_t           checksum;     //数据校验和
 	uint16_t           union_id;     //数据包的编号
-    uint64_t     	   cmd;          //命令
+    uint32_t     	   cmd;          //命令
     struct message     *msg;		 //
 }socket_pkg_t;
